==> notebooks/ag_impoundment_ira_usaspending_analysis/0_make_ag_awards_csv.ipynb <==
from pathlib import Path

import pandas as pd

pd.options.display.float_format = "{:,.0f}".format

data_dir: Path = Path().home() / "data" / "usaspending"

subdirs: list[Path] = [item for item in data_dir.iterdir() if item.is_dir()]

data_files: list[Path] = []

for subdir in subdirs:
    data_files += [
        i
        for i in subdir.iterdir()
        if i.is_file() and "All_Assistance_PrimeTransactions" in i.name
    ]

cols = [
    "assistance_transaction_unique_key",
    "assistance_award_unique_key",
    "award_id_fain",
    "modification_number",
    "federal_action_obligation",
    "total_obligated_amount",
    "total_outlayed_amount_for_overall_award",
    "indirect_cost_federal_share_amount",
    "non_federal_funding_amount",
    "total_non_federal_funding_amount",
    "face_value_of_loan",
    "total_face_value_of_loan",
    "total_loan_subsidy_cost",
    "generated_pragmatic_obligations",
    "action_date",
    "action_date_fiscal_year",
    "treasury_accounts_funding_this_award",
    "federal_accounts_funding_this_award",
    "object_classes_funding_this_award",
    "program_activities_funding_this_award",
    "recipient_uei",
    "recipient_name",
    "recipient_name_raw",
    "recipient_parent_uei",
    "recipient_parent_name",
    "recipient_parent_name_raw",
    "recipient_address_line_1",
    "recipient_address_line_2",
    "recipient_city_code",
    "recipient_city_name",
    "prime_award_transaction_recipient_county_fips_code",
    "recipient_county_name",
    "prime_award_transaction_recipient_state_fips_code",
    "recipient_state_code",
    "recipient_state_name",
    "recipient_zip_code",
    "recipient_zip_last_4_code",
    "prime_award_transaction_recipient_cd_original",
    "prime_award_transaction_recipient_cd_current",
    "primary_place_of_performance_scope",
    "primary_place_of_performance_code",
    "primary_place_of_performance_city_name",
    "prime_award_transaction_place_of_performance_county_fips_code",
    "primary_place_of_performance_county_name",
    "prime_award_transaction_place_of_performance_state_fips_code",
    "primary_place_of_performance_state_name",
    "primary_place_of_performance_zip_4",
    "prime_award_transaction_place_of_performance_cd_original",
    "prime_award_transaction_place_of_performance_cd_current",
    "primary_place_of_performance_foreign_location",
    "cfda_number",
    "cfda_title",
    "assistance_type_code",
    "assistance_type_description",
    "transaction_description",
    "prime_award_base_transaction_description",
    "business_funds_indicator_code",
    "business_funds_indicator_description",
    "business_types_code",
    "business_types_description",
    "correction_delete_indicator_code",
    "correction_delete_indicator_description",
    "action_type_code",
    "action_type_description",
    "record_type_code",
    "initial_report_date",
    "last_modified_date",
]

dfs: list[pd.DataFrame] = []
for i in data_files:
    df = pd.read_csv(i, usecols=cols)
    dfs.append(df)
df_out = pd.concat(dfs)
df_out.to_csv(data_dir / "ag_awards.csv", index=False)

* action_date_fiscal_year: fy in which the most recent change to the award occured; should be 2022 or later
* treasury_accounts_funding_this_award: tafs associated with the award; should be one of the IRA/IIJA tafs
