==> notebooks/ag_impoundment_ira_usaspending_analysis/2_make_ag_ira_awards_csv.ipynb <==
from pathlib import Path

import pandas as pd

pd.options.display.float_format = "{:,.0f}".format

data_dir: Path = Path().home() / "data" / "usaspending"

* action_date_fiscal_year: fy in which the most recent change to the award occured; should be 2022 or later
* treasury_accounts_funding_this_award: tafs associated with the award; should be one of the IRA/IIJA tafs

df = pd.read_csv(data_dir / "ag_awards_fy2022_fy2025.csv")

osec tafs
* 012-0115 22/31
* 012-0115 22/31
* 012-0115 22/31
* 012-0115 22/31
* 012-0115 22/31
* 012-0115 22/31
* 012-0115 22/28

fsa tafs
* 012-1124 22/31
* 012-1140 22/31

nrcs tafs
* "012-1000 22/31",
* "012-1000 22/31",
* "012-1004 23/31",
* "012-1004 24/31",
* "012-1004 25/31",
* "012-1004 23/31",
* "012-1004 24/31",
* "012-1004 25/31",
* "012-1004 26/31",
* "012-1004 23/31",
* "012-1004 24/31",
* "012-1004 25/31",
* "012-1004 26/31",
* "012-1004 23/31",
* "012-1004 24/31",
* "012-1004 25/31",
* "012-1004 26/31",
* "012-1004 23/31",
* "012-1004 24/31",
* "012-1004 25/31",
* "012-1004 26/31",

rcbs tafs
* 012-2073 22/31
* 012-1908 22/31
* 012-1908 23/31
* 012-1908 24/31
* 012-1908 25/31
* 012-1908 22/31
* 012-1908 23/31
* 012-1908 24/31
* 012-1908 25/31

rus tafs
* 012-1230 22/31
* 012-1230 22/31




tafs = [
    # osec
    "012-2022/2031-0115",
    "012-2024/2028-0115",
    # fsa
    "012-2022/2031-1124",
    "012-2022/2031-1140",
    # nifa
    "012-2022/2031-1500",
    # nrcs
    "012-2022/2031-1000",
    "012-2023/2031-1004",
    "012-2024/2031-1004",
    "012-2025/2031-1004",
    "012-2026/2031-1004",
    # rcbs
    "012-2022/2031-2073",
    "012-2022/2031-1908",
    "012-2023/2031-1908",
    "012-2024/2031-1908",
    "012-2025/2031-1908",
    "012-2026/2031-1908",
    # rus
    "012-2023/2031-1230",
]

mask = df["treasury_accounts_funding_this_award"].str.contains("|".join(tafs))
df = df.loc[mask]
df

df.loc[df["treasury_accounts_funding_this_award"].str.contains("1908")][
    "treasury_accounts_funding_this_award"
].value_counts()

df.to_csv(data_dir / "ag_ira_awards.csv", index=False)
