==> notebooks/ag_impoundment_ira_usaspending_analysis/3_make_ag_ira_withheld_csv.ipynb <==
from pathlib import Path

import pandas as pd

pd.options.display.float_format = "{:,.0f}".format

data_dir: Path = Path().home() / "data" / "usaspending"

df = pd.read_csv(data_dir / "ag_ira_awards.csv")

df["max_promised_amount"] = df[
    ["total_obligated_amount", "total_outlayed_amount_for_overall_award"]
].max(axis=1)
df["max_promised_amount"]

df.loc[df["treasury_accounts_funding_this_award"].str.contains("1908")][
    "treasury_accounts_funding_this_award"
].value_counts()

df["total_outlayed_amount_for_overall_award"] = df[
    "total_outlayed_amount_for_overall_award"
].fillna(0)

df["max_withheld_amount"] = (
    df["max_promised_amount"] - df["total_outlayed_amount_for_overall_award"]
)
df

df.loc[df["treasury_accounts_funding_this_award"].str.contains("1908")][
    "treasury_accounts_funding_this_award"
].value_counts()

df.loc[df["treasury_accounts_funding_this_award"].str.contains("1908")][
    ["treasury_accounts_funding_this_award", "max_withheld_amount"]
].value_counts()

df.loc[
    df["treasury_accounts_funding_this_award"].str.contains("1908")
    & (df["max_withheld_amount"] != 0)
][["treasury_accounts_funding_this_award", "max_withheld_amount"]]

df_withheld = df.loc[df.max_withheld_amount > 0]

df_withheld.loc[
    df_withheld["treasury_accounts_funding_this_award"].str.contains("1908")
]["treasury_accounts_funding_this_award"].value_counts()

df_withheld.to_csv(data_dir / "ag_ira_awards_withheld.csv", index=False)

df_withheld
