==> notebooks/ag_impoundment_ira_usaspending_analysis/5_reshape_withheld_awards_by_cd.ipynb <==
from pathlib import Path

import numpy as np
import pandas as pd

pd.options.display.float_format = "{:,.0f}".format

data_dir: Path = Path().home() / "data" / "usaspending"

df = pd.read_csv(data_dir / "ag_ira_awards_withheld.csv")

df.head()

df["prime_award_transaction_recipient_cd_current"] = df["prime_award_transaction_recipient_cd_current"].replace("-", "")
df["prime_award_transaction_recipient_cd_current"] = df["prime_award_transaction_recipient_cd_current"].replace("98", "00")
df["prime_award_transaction_recipient_cd_current"] = df["prime_award_transaction_recipient_cd_current"].replace("90", "Multiple")


eqip_tafs = [
    "012-2023/2031-1004",
    "012-2024/2031-1004",
    "012-2025/2031-1004",
    "012-2026/2031-1004",
]

cta_tafs = [
    "012-2022/2031-1000",
]

conditions = [
    df["cfda_title"] == "AGRICULTURAL CONSERVATION EASEMENT PROGRAM",
    df["cfda_title"] == "CONSERVATION RESERVE PROGRAM",
    df["cfda_title"] == "CONSERVATION STEWARDSHIP PROGRAM",
    df["cfda_title"] == "ENVIRONMENTAL QUALITY INCENTIVES PROGRAM",
    df["cfda_title"] == "REGIONAL CONSERVATION PARTNERSHIP PROGRAM",
    df["cfda_title"] == "RURAL ENERGY FOR AMERICA PROGRAM",
    df["cfda_title"] == "FROM LEARNING TO LEADING: CULTIVATING THE NEXT GENERATION OF DIVERSE FOOD AND AGRICULTURE PROFESSIONALS",

    (
        (df["cfda_title"] == "FARM AND RANCH LANDS PROTECTION PROGRAM")
        & (df["treasury_accounts_funding_this_award"].str.contains("|".join(eqip_tafs)))
    ),
    (
        (df["cfda_title"] == "FARM AND RANCH LANDS PROTECTION PROGRAM")
        & (df["treasury_accounts_funding_this_award"].str.contains("|".join(cta_tafs)))
    ),
    (
        (df["cfda_title"] == "SOIL AND WATER CONSERVATION")
        & (df["treasury_accounts_funding_this_award"].str.contains("|".join(eqip_tafs)))
    ),
    (
        (df["cfda_title"] == "SOIL AND WATER CONSERVATION")
        & (df["treasury_accounts_funding_this_award"].str.contains("|".join(cta_tafs)))
    ),
    (
        (df["cfda_title"] == "AGRICULTURAL AND RURAL ECONOMIC RESEARCH, COOPERATIVE AGREEMENTS AND COLLABORATIONS")
        & (df["treasury_accounts_funding_this_award"].str.contains("|".join(cta_tafs)))
    ),
]

values = [
    "Agricultural Conservation Easement Program",
    "Conservation Reserve Program",
    "Conservation Stewardship Program",
    "Environmental Quality Incentives Program",
    "Regional Conservation Partnership Program",
    "Rural Energy for America Program",
    "NIFA NEXTGEN",
    "Environmental Quality Incentives Program",
    "Conservation Technical Assistance",
    "Environmental Quality Incentives Program",
    "Conservation Technical Assistance",
    "Conservation Technical Assistance",
]

df["program_bucket"] = np.select(conditions, values, default=df["cfda_title"])

df["program_bucket"].value_counts()

df.to_csv(data_dir / "ag_ira_awards_withheld_program_buckets.csv", index=False)
